Conflicts:
	smart-user-client-impl/src/test/java/com/smartitengineering/clientrest/smartuserclientimpl/ComprehensiveClientTest.java

+++ smart-user-resource/src/main/java/com/smartitengineering/user/ws/resources/OrganizationPrivilegeResource.java
@@ -51,29 +51,38 @@ public class OrganizationPrivilegeResource extends AbstractResource {
 
     }
   }
-
-
   private String organizationUniqueShortName;
   private String privilegeName;
+  private Organization organization;
+  private Privilege privilege;
 
   public OrganizationPrivilegeResource(@PathParam("organizationUniqueShortName") String organizationUniqueShortName, @PathParam(
       "privilegeName") String privilegeName) {
     this.organizationUniqueShortName = organizationUniqueShortName;
     this.privilegeName = privilegeName;
+    organization = getOrganization();
+    privilege = getPrivilege();
   }
 
   @GET
   @Produces(MediaType.APPLICATION_JSON)
   @Path("/content")
   public Response getContent() {
-    ResponseBuilder responseBuilder = Response.ok(getPrivilege());
+    ResponseBuilder responseBuilder = Response.ok();
+    if (organization==null || privilege == null) {
+      return responseBuilder.status(Status.NOT_FOUND).build();
+    }
+    responseBuilder = Response.ok(getPrivilege());
     return responseBuilder.build();
   }
 
   @GET
   @Produces(MediaType.APPLICATION_ATOM_XML)
   public Response get() {
-    ResponseBuilder responseBuilder;
+    ResponseBuilder responseBuilder = Response.ok();
+    if (organization==null || privilege== null) {
+      return responseBuilder.status(Status.NOT_FOUND).build();
+    }
     try {
       responseBuilder = Response.status(Status.OK);
       Feed privilegeFeed = getPrivilegeFeed();
@@ -88,7 +97,10 @@ public class OrganizationPrivilegeResource extends AbstractResource {
 
   @DELETE
   public Response delete() {
-    ResponseBuilder responseBuilder;
+    ResponseBuilder responseBuilder = Response.ok();
+    if (organization==null || privilege == null) {
+      return responseBuilder.status(Status.NOT_FOUND).build();
+    }
     try {
       responseBuilder = Response.status(Status.OK);
       Services.getInstance().getPrivilegeService().delete(getPrivilege());
@@ -104,10 +116,13 @@ public class OrganizationPrivilegeResource extends AbstractResource {
   @Produces(MediaType.APPLICATION_ATOM_XML)
   @Consumes(MediaType.APPLICATION_JSON)
   public Response update(Privilege newPrivilege) {
-    ResponseBuilder responseBuilder;
+    ResponseBuilder responseBuilder = Response.ok();
+    if(organization==null || privilege==null){
+      return responseBuilder.status(Status.NOT_FOUND).build();
+    }
     try {
       responseBuilder = Response.status(Status.OK);
-      newPrivilege.setParentOrganization(getOrganization());
+      newPrivilege.setParentOrganization(organization);
       Services.getInstance().getPrivilegeService().update(newPrivilege);
     }
     catch (Exception ex) {
@@ -138,7 +153,6 @@ public class OrganizationPrivilegeResource extends AbstractResource {
     privilegeFeed.addLink(altLink);
 
     privilegeFeed.addLink(editLink);
-    privilegeFeed.addLink(altLink);
 
     return privilegeFeed;
   }
@@ -147,6 +161,9 @@ public class OrganizationPrivilegeResource extends AbstractResource {
   @Path("/delete")
   public Response deletePost() {
     ResponseBuilder responseBuilder = Response.ok();
+    if(organization==null || privilege==null){
+      return responseBuilder.status(Status.NOT_FOUND).build();
+    }
     try {
       Services.getInstance().getPrivilegeService().delete(getPrivilege());
     }
@@ -162,6 +179,9 @@ public class OrganizationPrivilegeResource extends AbstractResource {
   @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
   public Response updatePost(@HeaderParam("Content-type") String contentType, String message) {
     ResponseBuilder responseBuilder = Response.status(Status.SERVICE_UNAVAILABLE);
+    if(organization==null || privilege==null){
+      return responseBuilder.status(Status.NOT_FOUND).build();
+    }
 
     if (StringUtils.isBlank(message)) {
       responseBuilder = Response.status(Status.BAD_REQUEST);
@@ -188,8 +208,7 @@ public class OrganizationPrivilegeResource extends AbstractResource {
         ex.printStackTrace();
       }
     }
-    else {
-      contentType = contentType;
+    else {      
       isHtmlPost = false;
     }
